requirements.txt
fastapi
uvicorn
langchain
langchain-community
faiss-cpu
PyMuPDF
sentence-transformers
python-dotenv
pytest
